Conflicts:
	app/controllers/Groups.java
	app/dao/PersonReperibilityDayDao.java
	app/dao/ShiftDao.java
	app/manager/GroupManager.java
	conf/messages.it
	conf/permissions.drl

+++ app/controllers/Groups.java
@@ -1,3 +1,20 @@
+/*
+ * Copyright (C) 2021  Consiglio Nazionale delle Ricerche
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU Affero General Public License as
+ *     published by the Free Software Foundation, either version 3 of the
+ *     License, or (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU Affero General Public License for more details.
+ *
+ *     You should have received a copy of the GNU Affero General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 package controllers;
 
 import com.beust.jcommander.internal.Maps;
@@ -9,12 +26,16 @@ import dao.PersonDao;
 import dao.RoleDao;
 import dao.UsersRolesOfficesDao;
 import helpers.Web;
+import helpers.jpa.JpaReferenceBinder;
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 import javax.inject.Inject;
 import lombok.val;
 import lombok.extern.slf4j.Slf4j;
+import lombok.val;
 import manager.GroupManager;
 import models.GeneralSetting;
 import models.Office;
@@ -24,12 +45,17 @@ import models.User;
 import models.UsersRolesOffices;
 import models.dto.SeatSituationDto;
 import models.flows.Group;
+import org.testng.collections.Lists;
+import play.data.binding.As;
 import play.data.validation.Valid;
 import play.data.validation.Validation;
 import play.mvc.Controller;
 import play.mvc.With;
 import security.SecurityRules;
 
+/**
+ * Controller per la gestione dei gruppi.
+ */
 @Slf4j
 @With({Resecure.class})
 public class Groups extends Controller {
@@ -53,29 +79,41 @@ public class Groups extends Controller {
 
   /**
    * Metodo che crea il gruppo.
+   *
    * @param group il gruppo da creare
    * @param office la sede su cui crearlo
    */
-  public static void createGroup(@Valid Group group, Office office) {
-
+  public static void createGroup(
+      @Valid Group group, Office office,
+      @As(binder = JpaReferenceBinder.class)
+      Set<Person> people) {
+    log.info("affiltionaPeople = {}", people);
     if (Validation.hasErrors()) {
       response.status = 400;
-      render("@blank", office);
+      List<Person> peopleForGroups = personDao.byInstitute(office.institute);
+      log.info("Create groups errors = {}", validation.errorsMap());
+      render("@edit", group, office, peopleForGroups);
     }
     rules.checkIfPermitted(group.office);
     group.office = office;
+    final boolean isNew = group.id != null;        
     group.save();
-    log.debug("Salvato nuovo gruppo di lavoro: {} per la sede {}", group.name, group.office);
+    log.debug("Salvato gruppo di lavoro: {} per la sede {}", group.name, group.office);
+
+    groupManager.updatePeople(group, people);
+
     UsersRolesOffices uro = new UsersRolesOffices();
     groupManager.createManager(office, group, uro);
-    
-    flash.success("Nuovo gruppo  di lavoro %s salvato correttamente.",
-        group.name);
+    String message = isNew 
+        ? "Nuovo gruppo di lavoro %s salvato correttamente." 
+            : "Gruppo  di lavoro %s salvato correttamente.";
+    flash.success(message, group.name);
     showGroups(office.id);
   }
 
   /**
    * Metodo che cancella il gruppo.
+   *
    * @param groupId id del gruppo da cancellare
    */
   public static void deleteGroup(long groupId) {
@@ -84,20 +122,27 @@ public class Groups extends Controller {
     rules.checkIfPermitted(group.office);
 
     //elimino il ruolo di manager
-    if (!groupManager.deleteManager(group)) {
-      flash.error("Non esiste un manager associato al gruppo {}. "
-          + "Impossibile eliminarlo.", group.name);
-      showGroups(group.manager.office.id);
-    } 
-    //elimino il gruppo.
-    group.delete();
-    log.debug("Eliminato gruppo {}", group.name);
+    groupManager.deleteManager(group);
+
+    if (group.getPeople().isEmpty()) {
+      //Elimino eventuali vecchie associazioni
+      group.affiliations.stream().forEach(a -> a.delete());
+      //elimino il gruppo.
+      group.delete();
+      log.info("Eliminato gruppo {}", group.name);
+    } else {
+      group.endDate = LocalDate.now();
+      group.save();
+      log.info("Disattivato gruppo {}", group.name);
+    }
+
     flash.success(Web.msgDeleted(Group.class));
     showGroups(group.manager.office.id);
   }
 
   /**
    * Metodo che mostra i gruppi appartenenti a una sede.
+   *
    * @param officeId l'id della sede di cui vedere i gruppi
    */
   public static void showGroups(Long officeId) {
@@ -105,22 +150,31 @@ public class Groups extends Controller {
     notFoundIfNull(office);
     rules.checkIfPermitted(office);
     User user = Security.getUser().get();
-    List<Group> groups = null;
-    if (user.isSystemUser()) {
-      groups = groupDao.groupsByOffice(office, Optional.<Person>absent());
+    List<Group> groups = Lists.newArrayList();
+    if (uroDao.getUsersRolesOffices(user, roleDao.getRoleByName(Role.GROUP_MANAGER), office)
+        .isPresent()) {
+      groups = 
+          groupDao.groupsByOffice(office, Optional.fromNullable(user.person), Optional.of(true));
     }
-    if (user.hasRoles(Role.PERSONNEL_ADMIN)) {
-      groups = groupDao.groupsByOffice(office, Optional.<Person>absent());
+    if (user.isSystemUser() 
+        || uroDao.getUsersRolesOffices(user, roleDao.getRoleByName(Role.PERSONNEL_ADMIN), office)
+        .isPresent()) {
+      groups = groupDao.groupsByOffice(office, Optional.<Person>absent(), Optional.of(true));
     }
-    if (user.hasRoles(Role.GROUP_MANAGER)) {
-      groups = groupDao.groupsByOffice(office, Optional.fromNullable(user.person));
-    }
-     
-    render(groups, office);
+    val activeGroups = groups.stream().filter(g -> g.isActive())
+        .sorted((g1, g2) -> 
+            g1.getName().toLowerCase().compareTo(g2.getName().toLowerCase()))
+         .collect(Collectors.toList());
+    val disabledGroups = groups.stream().filter(g -> !g.isActive())
+        .sorted((g1, g2) -> 
+            g1.getName().toLowerCase().compareTo(g2.getName().toLowerCase()))
+        .collect(Collectors.toList());
+    render(activeGroups, disabledGroups, office);
   }
 
   /**
    * Metodo che permette la modifica del gruppo.
+   *
    * @param groupId id del gruppo da modificare
    */
   public static void edit(long groupId) {
@@ -134,6 +188,7 @@ public class Groups extends Controller {
 
   /**
    * Metodo che permette l'apertura della pagina di creazione del gruppo.
+   *
    * @param officeId l'id della sede su cui creare il gruppo
    */
   public static void blank(long officeId) {
@@ -147,7 +202,6 @@ public class Groups extends Controller {
     } else {
       peopleForGroups = personDao.byOffice(office);
     }
-    
     render("@edit", office, peopleForGroups);
   }
   
@@ -169,4 +223,8 @@ public class Groups extends Controller {
     Role role = roleDao.getRoleById(id);
     render(role);
   }
+
 }
+
+
+
